started splitting CUDA backend into the part common to HIP and CUDA and the part that is CUDA-specific (i.e. needs to be linked with CUDA)
